pre tidy up commit for breadcrumb
It is almost there. I need to pass though the actual name of the relevant IDGames page, but more importantly, there is a weird situation where the BC works as intended, but ONLY if the subsequent slug ID is numerically less than the current one. The tidy up may reveal where it is falling over. 

I also need to reduce the complexity of function calls. It's very inelegant and inefficient ATM.